pyproject.toml
# [build-system]
# requires = ["setuptools"]
# build-backend = "setuptools.build_meta"

[project]
name = "time_series_lab"
readme = "README.md"
dynamic = ["dependencies", "version"]

[project.scripts]
time-series-lab = "time_series_lab.__main__:main"

[project.entry-points."kedro.hooks"]

[project.optional-dependencies]
docs = [
    "docutils<0.18.0",
    "sphinx~=3.4.3",
    "sphinx_rtd_theme==0.5.1",
    "nbsphinx==0.8.1",
    "sphinx-autodoc-typehints==1.11.1",
    "sphinx_copybutton==0.3.1",
    "ipykernel>=5.3, <7.0",
    "Jinja2<3.1.0",
    "myst-parser~=0.17.2",
]

[tool.setuptools.dynamic]
dependencies = {file = "requirements.txt"}
version = {attr = "time_series_lab.__version__"}

[tool.setuptools.packages.find]
where = ["src"]
namespaces = false

[tool.kedro]
package_name = "time_series_lab"
project_name = "Time Series Lab"
kedro_init_version = "0.19.2"
tools = ['Linting', 'Testing', 'Custom Logging', 'Documentation', 'Data Structure', 'Kedro Viz']
example_pipeline = "False"
source_dir = "src"

[tool.pytest.ini_options]
addopts = """
--cov-report term-missing \
--cov src/time_series_lab -ra"""

[tool.coverage.report]
fail_under = 0
show_missing = true
exclude_lines = ["pragma: no cover", "raise NotImplementedError"]

[tool.ruff.format]
docstring-code-format = true

[tool.ruff]
line-length = 88
show-fixes = true
select = [
    "F",   # Pyflakes
    "W",   # pycodestyle
    "E",   # pycodestyle
    "I",   # isort
    "UP",  # pyupgrade
    "PL",  # Pylint
    "T201", # Print Statement
]
ignore = ["E501"]  # Ruff format takes care of line-too-long

[tool.poetry]
name = "time-series-lab"
version = "0.1.0"
description = ""
authors = ["dsthiago <thiagosd.2.18@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
kedro = "^0.19.2"
kedro-viz = "^7.1.0"
kedro-datasets = "^2.0.0"
pandas-datareader = "^0.10.0"
black = "^23.12.1"
yahoofinancials = "^1.20"
pyarrow = "^15.0.0"
fastparquet = "^2023.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
